requirements.txt
# Core Dependencies
numpy>=1.24.0
pandas<2.1.0
requests>=2.31.0
biopython>=1.83
rdkit>=2023.9.5
langgraph>=0.1.0

# API Integration
google-generativeai>=0.4.0

# Utilities
pyyaml>=6.0.1
python-dotenv>=1.0.0
tqdm>=4.66.1
jinja2>=3.1.3

# Development
pytest>=7.4.3
black>=24.3.0
mypy>=1.7.0
pylint>=3.0.2
